+++ environs/__init__.py
@@ -17,7 +17,7 @@ import marshmallow as ma
 from marshmallow.utils import _Missing
 from dotenv.main import load_dotenv, _walk_to_root
 
-__version__ = "8.0.0"
+__version__ = "8.1.0"
 __all__ = ["EnvError", "Env"]
 
 MARSHMALLOW_VERSION_INFO = tuple(int(part) for part in ma.__version__.split(".") if part.isdigit())